pyproject.toml
[project]
name = "lelamp-runtime"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "feetech-servo-sdk>=1.0.0",
    "langchain-core>=1.0.3",
    "langchain-openai>=1.0.2",
    "langgraph>=0.2.0",
    "lerobot>=0.3.3",
    "livekit-agents[openai]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "numpy>=2.2.6",
    "pvporcupine>=3.0.5",
    "pvrecorder>=1.2.7",
    "pyarrow==20.0.0",
    "pyaudio>=0.2.14",
    "python-dotenv",
    "sounddevice>=0.5.2",
]

[tool.uv.sources]
lerobot = { git = "https://github.com/huggingface/lerobot" }

[project.optional-dependencies]
hardware = [
    "adafruit-circuitpython-neopixel",
    "rpi-ws281x",
]
